+++ app/client/src/ee/pages/AdminSettings/WithSuperUserHoc.tsx
@@ -1,0 +1,3 @@
+export * from "ce/pages/AdminSettings/WithSuperUserHoc";
+import { default as CE_WithSuperUserHOC } from "ce/pages/AdminSettings/WithSuperUserHoc";
+export default CE_WithSuperUserHOC;
